backend/mobile_take_home_backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
SQLAlchemy==2.0.34
pydantic==2.8.2
python-multipart==0.0.12
Pillow==10.4.0
